pyproject.toml
[project]
name = "openphone-python"
dynamic = ["version"]
description = "Python SDK for the OpenPhone API"
readme = "README.md"
authors = [
    { name = "plug-city", email = "30635146+plug-city@users.noreply.github.com" }
]
requires-python = ">=3.13"
dependencies = [
    "requests>=2.25.0",
    "phonenumbers>=8.12.0",
    "python-dateutil>=2.8.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=6.0.0",
    "responses>=0.18.0",
    "black>=22.0.0",
    "flake8>=4.0.0",
    "mypy>=0.900",
    "pre-commit>=2.20.0",
]

[project.urls]
Homepage = "https://github.com/plug-city/openphone-python"
Documentation = "https://github.com/plug-city/openphone-python"
Repository = "https://github.com/plug-city/openphone-python"

[tool.black]
line-length = 88
target-version = ['py311']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
check_untyped_defs = true
strict_optional = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--strict-markers",
    "--strict-config",
    "-ra",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.version]
path = "src/openphone_python/_version.py"

[dependency-groups]
dev = [
    "black>=25.1.0",
    "flake8>=7.3.0",
    "mypy>=1.17.1",
    "pre-commit>=4.2.0",
    "pytest>=8.4.1",
    "responses>=0.25.7",
]
